+++ CHANGELOG.md
@@ -11,6 +11,10 @@ Release: dd.mm.yyyy
 - Forward `console.log(...)` to `Phoenix.log(...)` for convenience.
 - Log stack trace when passing an error to `Phoenix.log(...)` ([#181](https://github.com/kasper/phoenix/issues/181)).
 
+### Improvements
+
+- Launch Console app when clicking a notification for an exception.
+
 2.6.7
 -----